+++ typescript/web/src/components/labeling-tool/openlayers-map/draw-interaction/index.tsx
@@ -1,13 +1,17 @@
 import { MutableRefObject } from "react";
 import { useRouter } from "next/router";
+import { Vector as OlSourceVector } from "ol/source";
+import { Geometry } from "ol/geom";
 import { useLabelingStore, Tools } from "../../../../connectors/labeling-state";
-import { DrawIogInteraction } from "../iog";
+import { DrawIogInteraction } from "../iog/draw";
 import { DrawBoundingBoxAndPolygonInteraction } from "./bounding-box-polygon";
 
 export const DrawInteraction = ({
   iogSpinnerRef,
+  sourceVectorLabelsRef,
 }: {
   iogSpinnerRef: MutableRefObject<HTMLDivElement | null>;
+  sourceVectorLabelsRef: MutableRefObject<OlSourceVector<Geometry> | null>;
 }) => {
   const { imageId } = useRouter()?.query;
 
@@ -20,7 +24,11 @@ export const DrawInteraction = ({
     return null;
   }
   return selectedTool === Tools.IOG ? (
-    <DrawIogInteraction imageId={imageId} iogSpinnerRef={iogSpinnerRef} />
+    <DrawIogInteraction
+      imageId={imageId}
+      iogSpinnerRef={iogSpinnerRef}
+      sourceVectorLabelsRef={sourceVectorLabelsRef}
+    />
   ) : (
     <DrawBoundingBoxAndPolygonInteraction imageId={imageId} />
   );
